remove all references to fivetransdk in lib
