+++ cmd/internal/server/server.go
@@ -163,7 +163,7 @@ func (c *connectorServer) Update(request *fivetran_sdk.UpdateRequest, server fiv
 		return status.Errorf(codes.InvalidArgument, "unable to list shards for this database : %q", err)
 	}
 	var state *lib.SyncState
-	state, err = lib.StateFromRequest(request, *psc, shards, *schema)
+	state, err = handlers.StateFromRequest(request, *psc, shards, *schema)
 	if err != nil {
 		return status.Error(codes.InvalidArgument, fmt.Sprintf("request did not contain a valid stateJson : %q", err))
 	}
